--- src/password-input/README.md
@@ -1,11 +1,0 @@
-# @dojo/widgets/password-input
-
-Dojo's `PasswordInput` widget provides a basic password input widget with an optional label.
-
-
-## Features
-
-
-### Accessibility Features
-
-
